refactor: create index.ts files for all APIes
