+++ packages/client/src/api/authAPI/index.ts
@@ -1,0 +1,2 @@
+export { authAPI } from './authAPI';
+export { type SigninRequestData, type SignupRequestData } from './typings';
