+++ packages/client/src/api/contactAPI/index.ts
@@ -1,0 +1,1 @@
+export { contactAPI } from './contactAPI';
