+++ packages/client/src/api/forumAPI/index.ts
@@ -1,0 +1,1 @@
+export { forumAPI } from './forumAPI';
