+++ packages/client/src/api/leaderboardAPI/index.ts
@@ -1,0 +1,2 @@
+export { leaderboardAPI } from './leaderboardAPI';
+export type { GetLeaderboardResponseData, LeaderboardRecord, LeaderboardRecordData } from './typings';
