+++ packages/client/src/api/oauthAPI/index.ts
@@ -1,0 +1,1 @@
+export { oauthAPI } from './oauthAPI';
