+++ packages/client/src/api/themizationAPI/index.ts
@@ -1,0 +1,1 @@
+export { themizationAPI } from './themizationAPI';
