+++ packages/client/src/api/userAPI/index.ts
@@ -1,0 +1,2 @@
+export type { UpdatePasswordRequestData, UpdateProfileAvatarRequestData, UpdateProfileRequestData } from './typings';
+export { userAPI } from './userAPI';
